requirements.txt
pytest>=2.7
uiautomator>=0.2.4
Pillow>=3.0.0
future>=0.15.2
mock>=1.3.0
